KFL-AQGen-AI-ezzyoung_rag_agentized/requirements.txt
rank_bm25
faiss-cpu
pandas
python-dotenv
pydantic<3
openai>=1.30
tiktoken>=0.5
langchain-core>=0.2,<0.3
langchain>=0.2,<0.3
langchain-community>=0.2,<0.3
langchain-openai>=0.1,<0.2
langgraph>=0.2,<0.3
